[llvm] [refactor] (Decomp of #5251 6/n) Let ModuleToFunctionConverter support multiple modules
