+++ taichi/codegen/codegen.h
@@ -1,6 +1,7 @@
 // Driver class for kernel code generators.
 
 #pragma once
+#include <taichi/runtime/llvm/llvm_runtime_executor.h>
 #include "taichi/ir/ir.h"
 #include "taichi/program/program.h"
 #ifdef TI_WITH_LLVM
@@ -36,4 +37,26 @@ class KernelCodeGen {
 #endif
 };
 
+#ifdef TI_WITH_LLVM
+
+class ModuleToFunctionConverter {
+ public:
+  explicit ModuleToFunctionConverter(TaichiLLVMContext *tlctx,
+                                     LlvmRuntimeExecutor *program);
+
+  virtual ~ModuleToFunctionConverter() = default;
+
+  virtual FunctionType convert(const std::string &kernel_name,
+                               const std::vector<LlvmLaunchArgInfo> &args,
+                               std::vector<LLVMCompiledData> &&data) const = 0;
+
+  virtual FunctionType convert(const Kernel *kernel,
+                               std::vector<LLVMCompiledData> &&data) const;
+
+ protected:
+  TaichiLLVMContext *tlctx_{nullptr};
+  LlvmRuntimeExecutor *executor_{nullptr};
+};
+
+#endif
 TLANG_NAMESPACE_END
