+++ taichi/codegen/cpu/codegen_cpu.h
@@ -10,7 +10,7 @@
 TLANG_NAMESPACE_BEGIN
 
 class CodeGenCPU : public KernelCodeGen {
- public:
+public:
   CodeGenCPU(Kernel *kernel, IRNode *ir = nullptr) : KernelCodeGen(kernel, ir) {
   }
 
@@ -22,5 +22,23 @@ class CodeGenCPU : public KernelCodeGen {
 
   FunctionType codegen() override;
 };
+
+#ifdef TI_WITH_LLVM
+
+class CPUModuleToFunctionConverter : public ModuleToFunctionConverter {
+ public:
+  explicit CPUModuleToFunctionConverter(TaichiLLVMContext *tlctx,
+                                        LlvmRuntimeExecutor *executor)
+      : ModuleToFunctionConverter(tlctx, executor) {
+  }
+
+  using ModuleToFunctionConverter::convert;
+
+  FunctionType convert(const std::string &kernel_name,
+                       const std::vector<LlvmLaunchArgInfo> &args,
+                       std::vector<LLVMCompiledData> &&data) const override;
+};
+
+#endif
 
 TLANG_NAMESPACE_END
